+++ frontend/src/components/prompts/CreateApi.vue
@@ -8,6 +8,7 @@
       <!-- API Key Name Input -->
       <p>{{ $t('api.keyName') }}</p>
       <input
+        v-focus
         class="input input--block"
         type="text"
         v-model.trim="apiName"