+++ CHANGELOG.md
@@ -4,11 +4,22 @@ All notable changes to this project will be documented in this file. For commit 
 
 ## v0.7.15-beta
 
+ **New Features**:
+ - added userDefault `disableViewingExt`. The new properties apply to all files, not just office.
+ - code blocks in markdown viewer have line numbers and each line is highlightable
+
+ **Notes**:
+ - replaced `disableOfficePreviewExt` with more generally applicable `disablePreviewExt` to disable preview for any specific file type.
+ - more tooltip descriptions for settings options
+
  **BugFixes**:
  - fix chinese and other language error (#972, #969)
  - fix docker dockerfile for `docker run` (#973)
  - fix double slash href on single source (#968)
  - fix sources named "files" or "share" issue (#949, #574)
+ - focus input field on popups (#976)
+ - hopeful fix for size calculation (#982)
+ - edit button is not working on .md files (#983)
 
 ## v0.7.14-beta